+++ src/cli/context.js
@@ -1,140 +1,18 @@
 "use strict";
 
 const path = require("path");
-const minimist = require("minimist");
 const fs = require("fs");
 const globby = require("globby");
 const chalk = require("chalk");
 const readline = require("readline");
 
 const util = require("./util");
-const normalizer = require("../main/options-normalizer");
 const prettier = require("../../index");
 const cleanAST = require("../common/clean-ast").cleanAST;
 const errors = require("../common/errors");
-const constant = require("./constant");
 const optionsModule = require("../main/options");
 const optionsNormalizer = require("../main/options-normalizer");
 const thirdParty = require("../common/third-party");
-const getSupportInfo = require("../common/support").getSupportInfo;
-const commonUtil = require("../common/util");
-
-// context
-
-function updateContextOptions(context, plugins) {
-  const supportOptions = getSupportInfo(null, {
-    showDeprecated: true,
-    showUnreleased: true,
-    showInternal: true,
-    plugins
-  }).options;
-
-  const detailedOptionMap = util.normalizeDetailedOptionMap(
-    Object.assign(
-      {},
-      util.createDetailedOptionMap(supportOptions),
-      constant.options
-    )
-  );
-
-  const detailedOptions = commonUtil.arrayify(detailedOptionMap, "name");
-
-  const apiDefaultOptions = supportOptions
-    .filter(optionInfo => !optionInfo.deprecated)
-    .reduce(
-      (reduced, optionInfo) =>
-        Object.assign(reduced, { [optionInfo.name]: optionInfo.default }),
-      Object.assign({}, optionsModule.hiddenDefaults)
-    );
-
-  context.supportOptions = supportOptions;
-  context.detailedOptions = detailedOptions;
-  context.detailedOptionMap = detailedOptionMap;
-  context.apiDefaultOptions = apiDefaultOptions;
-}
-
-function pushContextPlugins(context, plugins) {
-  context._supportOptions = context.supportOptions;
-  context._detailedOptions = context.detailedOptions;
-  context._detailedOptionMap = context.detailedOptionMap;
-  context._apiDefaultOptions = context.apiDefaultOptions;
-  updateContextOptions(context, plugins);
-}
-
-function popContextPlugins(context) {
-  context.supportOptions = context._supportOptions;
-  context.detailedOptions = context._detailedOptions;
-  context.detailedOptionMap = context._detailedOptionMap;
-  context.apiDefaultOptions = context._apiDefaultOptions;
-}
-
-function updateContextArgv(context, plugins) {
-  pushContextPlugins(context, plugins);
-
-  const minimistOptions = util.createMinimistOptions(context.detailedOptions);
-  const argv = minimist(context.args, minimistOptions);
-
-  context.argv = argv;
-  context.filePatterns = argv["_"];
-}
-
-function normalizeContextArgv(context, keys) {
-  const detailedOptions = !keys
-    ? context.detailedOptions
-    : context.detailedOptions.filter(
-        option => keys.indexOf(option.name) !== -1
-      );
-  const argv = !keys ? context.argv : util.pick(context.argv, keys);
-
-  context.argv = normalizer.normalizeCliOptions(argv, detailedOptions, {
-    logger: context.logger
-  });
-}
-
-//
-
-function listDifferent(context, input, options, filename) {
-  if (!context.argv["list-different"]) {
-    return;
-  }
-
-  options = Object.assign({}, options, { filepath: filename });
-
-  if (!prettier.check(input, options)) {
-    if (!context.argv["write"]) {
-      context.logger.log(filename);
-    }
-    process.exitCode = 1;
-  }
-
-  return true;
-}
-
-function handleError(context, filename, error) {
-  const isParseError = Boolean(error && error.loc);
-  const isValidationError = /Validation Error/.test(error && error.message);
-
-  // For parse errors and validation errors, we only want to show the error
-  // message formatted in a nice way. `String(error)` takes care of that. Other
-  // (unexpected) errors are passed as-is as a separate argument to
-  // `console.error`. That includes the stack trace (if any), and shows a nice
-  // `util.inspect` of throws things that aren't `Error` objects. (The Flow
-  // parser has mistakenly thrown arrays sometimes.)
-  if (isParseError) {
-    context.logger.error(`${filename}: ${String(error)}`);
-  } else if (isValidationError || error instanceof errors.ConfigError) {
-    context.logger.error(String(error));
-    // If validation fails for one file, it will fail for all of them.
-    process.exit(1);
-  } else if (error instanceof errors.DebugError) {
-    context.logger.error(`${filename}: ${error.message}`);
-  } else {
-    context.logger.error(filename + ": " + (error.stack || error));
-  }
-
-  // Don't exit the process if one file failed
-  process.exitCode = 2;
-}
 
 /**
  * @property supportOptions
@@ -146,17 +24,17 @@ class Context {
   constructor(args) {
     this.args = args;
 
-    updateContextArgv(this);
-    normalizeContextArgv(this, ["loglevel", "plugin"]);
+    util.updateContextArgv(this);
+    util.normalizeContextArgv(this, ["loglevel", "plugin"]);
 
     this.logger = util.createLogger(this.argv["loglevel"]);
 
-    updateContextArgv(this, this.argv["plugin"]);
+    util.updateContextArgv(this, this.argv["plugin"]);
   }
 
   init() {
     // split into 2 step so that we could wrap this in a `try..catch` in cli/index.js
-    normalizeContextArgv(this);
+    util.normalizeContextArgv(this);
   }
 
   formatStdin() {
@@ -175,14 +53,14 @@ class Context {
 
       const options = this.getOptionsForFile(filepath);
 
-      if (listDifferent(this, input, options, "(stdin)")) {
+      if (util.listDifferent(this, input, options, "(stdin)")) {
         return;
       }
 
       try {
         util.writeOutput(this.format(input, options), options);
       } catch (error) {
-        handleError(this, "stdin", error);
+        util.handleError(this, "stdin", error);
       }
     });
   }
@@ -221,7 +99,7 @@ class Context {
         return;
       }
 
-      listDifferent(this, input, options, filename);
+      util.listDifferent(this, input, options, filename);
 
       const start = Date.now();
 
@@ -238,7 +116,7 @@ class Context {
         // Add newline to split errors from filename line.
         process.stdout.write("\n");
 
-        handleError(this, filename, error);
+        util.handleError(this, filename, error);
         return;
       }
 
@@ -371,7 +249,7 @@ class Context {
 
     const hasPlugins = options && options.plugins;
     if (hasPlugins) {
-      pushContextPlugins(this, options.plugins);
+      util.pushContextPlugins(this, options.plugins);
     }
 
     const appliedOptions = Object.assign(
@@ -392,7 +270,7 @@ class Context {
     );
 
     if (hasPlugins) {
-      popContextPlugins(this);
+      util.popContextPlugins(this);
     }
 
     return appliedOptions;
